# perf(clean): improve the peformace of the clean subpackage