… dataprep/clean/clean_ip.py
… dataprep/clean/clean_lat_long.py
… dataprep/clean/clean_url.py
… dataprep/tests/clean/test_clean_ip.py
… dataprep/tests/clean/test_clean_url.py
… docs/source/user_guide/clean/clean_lat_long.ipynb
… docs/source/user_guide/clean/clean_url.ipynb