# Description

The main efficiency gain is from using series-level operations rather than a DataFrame row-wise apply. Another improvements is not using a global dictionary for the stats.

Some experimental results run on my local machine

| Function | Row count | Runtime with optimizations | Runtime without optimizations| Dataset | Column(s) |
| --- | --- | --- | --- | --- | --- |
| clean_lat_long | 50000 | **1.4s** | **60s** | `pd.read_csv("https://raw.githubusercontent.com/imdevskp/covid_19_jhu_data_web_scrap_and_cleaning/master/covid_19_clean_complete.csv")` | Lat, Long |
| clean_lat_long | 1 million | **28s** |  | `pd.read_csv("https://raw.githubusercontent.com/imdevskp/covid_19_jhu_data_web_scrap_and_cleaning/master/covid_19_clean_complete.csv")` | Lat, Long |
| clean_country | 50000 | **1s** | **61s** | `pd.read_csv("https://raw.githubusercontent.com/imdevskp/covid_19_jhu_data_web_scrap_and_cleaning/master/covid_19_clean_complete.csv")` | Country/Region |
| clean_country | 1 million | **19s** |  | `pd.read_csv("https://raw.githubusercontent.com/imdevskp/covid_19_jhu_data_web_scrap_and_cleaning/master/covid_19_clean_complete.csv")` | Country/Region |
| clean_phone | 50000 | **0.8s** | **55s** | https://catalogue.data.gov.bc.ca/dataset/bc-indigenous-business-listings/resource/1b9d73d0-62a7-485f-a177-55e46701073a | Phone |
| clean_phone | 1 million | **16.7s** |  | https://catalogue.data.gov.bc.ca/dataset/bc-indigenous-business-listings/resource/1b9d73d0-62a7-485f-a177-55e46701073a` | Phone |
| clean_url | 50000 | **1.1s** | **55s** | https://catalogue.data.gov.bc.ca/dataset/bc-indigenous-business-listings/resource/1b9d73d0-62a7-485f-a177-55e46701073a | Web Site |
| clean_url | 1 million | **24s** |  | https://catalogue.data.gov.bc.ca/dataset/bc-indigenous-business-listings/resource/1b9d73d0-62a7-485f-a177-55e46701073a` | Web Site |

# How Has This Been Tested?

Unit tests

# Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have already squashed the commits and make the commit message conform to the project standard.
- [x] I have already marked the commit with "BREAKING CHANGE" or "Fixes #" if needed.
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged and published in downstream modules