@brandonlockhart 2021-01-03
Good idea @jnwang. I added the progress bar from EDA, but it seems slow to update at the beginning (stays at 0%). It might be because the cleaning operation takes a long time and the progress bar does not update until it's done.

![ezgif com-gif-maker](https://user-images.githubusercontent.com/25874021/103486660-3dba3e80-4db4-11eb-8138-32b7eae59cf8.gif)