@jnwang 2021-01-04
What will the progress bar look like when cleaning 1 million urls (24s in total)?