@brandonlockhart 2021-01-04
I don't know what's causing the delay before the progress bar appears and starts updating, but I think it's better to have it than not.

https://user-images.githubusercontent.com/25874021/103494436-b4266300-4deb-11eb-832f-7a58d8050bb3.mov